fix lgtm alerts (getsentry/sentry-android#219)

+++ sentry-core/src/main/java/io/sentry/core/transport/IBackOffIntervalStrategy.java
@@ -15,5 +15,5 @@ public interface IBackOffIntervalStrategy {
    * @param attempt the number of attempts made so far
    * @return the suggested delay in milliseconds
    */
-  long nextDelayMillis(int attempt);
+  long nextDelayMillis(long attempt);
 }
